% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollucheck_run.R
\name{pollucheck_run}
\alias{pollucheck_run}
\title{pollucheck: Explore Open-Source Air-Quality Data}
\usage{
pollucheck_run()
}
\description{
pollucheck: Explore Open-Source Air-Quality Data
}
\examples{
\dontrun{
library(pollucheck)
pollucheck::pollucheck_run()
}

}
\keyword{pollucheck}
