% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollucheck-package.R
\docType{package}
\name{pollucheck-package}
\alias{pollucheck}
\alias{pollucheck-package}
\title{pollucheck: Explore Open-Source Air-Pollution Data}
\description{
This app helps to analyze and visualize open-source air
    pollution data available.  An example dataset is already loaded to help
    you walk through all the features of the app.  If you need help with
    downloading your own, check the ReadMe file on github.  This app can
    process all parameters except Wind Direction which when downloaded at
    1 hour is processed correctly, any other time-resolution is used then
    the app will not be process wind direction correctly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adithirgis/OpenSourceAirQualityApp}
  \item Report bugs at \url{https://github.com/adithirgis/OpenSourceAirQualityApp/issues}
}

}
\author{
\strong{Maintainer}: Adithi R. Upadhya \email{adithi@ilklabs.com} [copyright holder]

Authors:
\itemize{
  \item Meenakshi Kushwaha \email{meenakshi@ilklabs.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Pratyush Agrawal \email{pratyush@ilklabs.com} [data contributor]
  \item Sreekanth Vakacherla \email{sree.hcu@gmail.com} [author of introduction, etc., reviewer]
}

}
\keyword{internal}
