% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollucheck_run.R
\name{pollucheck_run}
\alias{pollucheck_run}
\title{pollucheck: Explore Open-Source Air-Quality Data}
\usage{
pollucheck_run()
}
\description{
This app helps to analyze and visualize open source air
quality data available.  An example dataset is already loaded to help
you walk through all the features of the app.  If you need help with
downloading your own, check the ReadMe file on github.  This app can
process all parameters except Wind Direction which when downloaded at
1 hour is processed correctly, any other time-resolution is used then
the app will not be process wind direction correctly.
}
\examples{
\dontrun{
library(pollucheck)
pollucheck::pollucheck_run()
}

}
\keyword{pollucheck}
