% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollucheck_run.R
\name{pollucheck_run}
\alias{pollucheck_run}
\title{pollucheck: Explore Open-Source Air Quality Data}
\usage{
pollucheck_run()
}
\description{
Helps to analyze and visualize open source air quality data available
online which are - Central Pollution Control Board (India), OpenAQ, AirNow.
An example dataset is already loaded for user to walk through all the features of the app.
If any assistance is required with downloading dataset, check the README file on GitHub
This app can process all parameters except wind direction which when downloaded at
1 hour is processed correctly. When wind direction any other time resolution
(e.g. - daily average) is not averaged correctly as it is a vector and
cannot be averaged using simple arithmetic mean.
}
\examples{
if(interactive()){
library(pollucheck)
pollucheck_run()
}

}
\keyword{pollucheck}
