% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollucheck_run.R
\name{pollucheck_run}
\alias{pollucheck_run}
\title{pollucheck: Explore Open-Source Air-Quality Data}
\usage{
pollucheck_run()
}
\description{
Helps to analyze and visualize open source air quality data available
online which are - Central Pollution Control Board (India), OpenAQ, AirNow.
An example dataset is already loaded for user to walk through all the features of the app.
If any assistance is required with downloading dataset, check the ReadMe file on github.
This app can process all parameters except Wind Direction which when downloaded at
1 hour is processed correctly, and if any other time-resolution is used then
the app will not process wind direction correctly.
}
\examples{
if(interactive()){
library(pollucheck)
pollucheck_run()
}

}
\keyword{pollucheck}
